Changed the RTL logic back to using its own global internally. This code is still non-working.. well, at least it isn't tested.
I really want to modify the RTL so that it is completely copy on write. The
trick with that is all packet work must be done on the directly imported
version. I don't want to start soaking up buffers.

I have an idea about all this though. We'll see where my limited knowledge
of unions takes us.