A random commit to make it seem as though I'm doing work.
I've decided to continue working on the debug driver and will continue until
it is finished.

Screw this making it professional until I make it work. I like the feeling
of progress, again.

UDP checksums and new checksum architecture is in. I think the designers of
TCP/IP were completely unjustified in their breaking of the layering
paradigm - but it's not like I can make it changed.

I bet its even worse in IPv6.