UDP/IP is now completely working. I handled the 16-bit alignment problem. Pad with nulls baby.
I'd like to take this moment to plug "Internetworking with TCP/IP: Volume
II". This is the volume I don't have, but the source code for which was
immensely helpful. On only this. The rest I of the time I spent whacking my
head against the abstractness of Volume I.

Of course, that's what the point of Volume I is, but I can still complain.
Right?